API/requirements.txt
aniso8601==9.0.1
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
PyJWT==2.3.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.35
Werkzeug==2.1.1
zipp==3.8.0
